+++ bundle/abusefilter.ts
@@ -1,0 +1,4 @@
+import {registerAbuseFilterCore} from '../src/index';
+export {CodeMirror6} from '../src/index';
+
+registerAbuseFilterCore();