+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## 3.10.0
 
-*2026-01-17*
+*2026-01-19*
 
 **Added**
 
 - [CodeFolding](./README.md#codefolding) keyboard shortcuts now work with sections and tables in the MediaWiki mode
 - New optional argument for [`registerMediaWiki`](./README.md#mediawiki) to specify the article path of a MediaWiki site
+- Language support for [Extension:AbuseFilter](https://www.mediawiki.org/wiki/Extension:AbuseFilter)
 
 **Fixed**